refactor: move ui-message modules
