+++ apps/meteor/client/hooks/roomActions/useAppsRoomStarActions.tsx
@@ -5,8 +5,8 @@ import type { RoomToolboxActionConfig } from '@rocket.chat/ui-contexts';
 import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import { UiKitTriggerTimeoutError } from '../../../app/ui-message/client/UiKitTriggerTimeoutError';
 import { Utilities } from '../../../ee/lib/misc/Utilities';
+import { UiKitTriggerTimeoutError } from '../../lib/errors/UiKitTriggerTimeoutError';
 import { useUiKitActionManager } from '../../uikit/hooks/useUiKitActionManager';
 import { useRoom } from '../../views/room/contexts/RoomContext';
 import { useAppActionButtons } from '../useAppActionButtons';
