+++ apps/meteor/client/hooks/useMessageboxAppsActionButtons.ts
@@ -4,8 +4,8 @@ import { useTranslation } from 'react-i18next';
 
 import { useAppActionButtons, getIdForActionButton } from './useAppActionButtons';
 import { useApplyButtonFilters } from './useApplyButtonFilters';
-import { UiKitTriggerTimeoutError } from '../../app/ui-message/client/UiKitTriggerTimeoutError';
 import { Utilities } from '../../ee/lib/misc/Utilities';
+import { UiKitTriggerTimeoutError } from '../lib/errors/UiKitTriggerTimeoutError';
 import type { MessageBoxAction } from '../lib/messageBox';
 import { useUiKitActionManager } from '../uikit/hooks/useUiKitActionManager';
