+++ apps/meteor/client/hooks/useUserDropdownAppsActionButtons.ts
@@ -6,8 +6,8 @@ import { useTranslation } from 'react-i18next';
 
 import { useAppActionButtons } from './useAppActionButtons';
 import { useApplyButtonAuthFilter } from './useApplyButtonFilters';
-import { UiKitTriggerTimeoutError } from '../../app/ui-message/client/UiKitTriggerTimeoutError';
 import { Utilities } from '../../ee/lib/misc/Utilities';
+import { UiKitTriggerTimeoutError } from '../lib/errors/UiKitTriggerTimeoutError';
 import { useUiKitActionManager } from '../uikit/hooks/useUiKitActionManager';
 
 export const useUserDropdownAppsActionButtons = () => {
