+++ apps/meteor/client/providers/ActionManagerProvider.tsx
@@ -1,11 +1,11 @@
 import { ActionManagerContext, useRouter } from '@rocket.chat/ui-contexts';
 import type { ReactNode } from 'react';
 
-import { ActionManager } from '../../app/ui-message/client/ActionManager';
 import { useAppActionButtons } from '../hooks/useAppActionButtons';
 import { useAppSlashCommands } from '../hooks/useAppSlashCommands';
 import { useAppUiKitInteraction } from '../hooks/useAppUiKitInteraction';
 import { useTranslationsForApps } from '../hooks/useTranslationsForApps';
+import { ActionManager } from '../lib/ActionManager';
 import { useInstance } from '../views/room/providers/hooks/useInstance';
 
 export type ActionManagerProviderProps = {
